turn data_to_sample into an executable submodule; refactor the centroid model's setup to use the new submodule
